Protection (acces au page que si connecté)

+++ src/Controller/BoissonController.php
@@ -22,6 +22,9 @@ class BoissonController extends AbstractController
     #[Route('/newBoisson', name: 'Boisson_new', methods: ['GET', 'POST'])]
     public function new(Request $request, BoissonRepository $BoissonRepository, FileUploader $fileUploader): Response
     {
+        if (!$this->isGranted('IS_AUTHENTICATED_FULLY')) {
+            return $this->redirectToRoute('app_accueil');
+        } else {
         $boisson = new Boisson();
         $createBoissonform = $this->createForm(BoissonType::class, $boisson);
         $createBoissonform->handleRequest($request);
@@ -43,7 +46,7 @@ class BoissonController extends AbstractController
             'boisson' => $boisson,
             'Boisson' => $createBoissonform->createView(),
         ]);
-    }
+    }}
 
 
     /////////////////////////////////////////...update...\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
@@ -51,41 +54,48 @@ class BoissonController extends AbstractController
     #[Route('/{id}/editBoisson', name: 'menu_edit_Boisson', methods: ['GET', 'POST'])]
     public function edit(Request $request, Boisson $boisson, BoissonRepository $boissonRepository, FileUploader $fileUploader): Response
     {
-        $modifierBoissonForm = $this->createForm(BoissonType::class, $boisson);
-        $modifierBoissonForm->handleRequest($request);
+        if (!$this->isGranted('IS_AUTHENTICATED_FULLY')) {
+            return $this->redirectToRoute('app_accueil');
+        } else {
+            $modifierBoissonForm = $this->createForm(BoissonType::class, $boisson);
+            $modifierBoissonForm->handleRequest($request);
 
-        if ($modifierBoissonForm->isSubmitted() && $modifierBoissonForm->isValid()) {
+            if ($modifierBoissonForm->isSubmitted() && $modifierBoissonForm->isValid()) {
 
-            //################# image ###################\\
-            /** @var UploadedFile $brochureFile */
-            $brochureFile = $modifierBoissonForm->get('brochure')->getData();
-            if ($brochureFile) {
-                $brochureFileName = $fileUploader->upload($brochureFile);
-                $boisson->setBrochureFilename($brochureFileName);
+                //################# image ###################\\
+                /** @var UploadedFile $brochureFile */
+                $brochureFile = $modifierBoissonForm->get('brochure')->getData();
+                if ($brochureFile) {
+                    $brochureFileName = $fileUploader->upload($brochureFile);
+                    $boisson->setBrochureFilename($brochureFileName);
+                }
+
+                $boissonRepository->add($boisson, true);
+                return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
             }
 
-            $boissonRepository->add($boisson, true);
-            return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
+            $brochure =$boisson->getBrochureFilename();
+            $ID = $boisson->getId();
+            return $this->render('Boisson/editBoisson.html.twig', [
+                'ID'=>$ID,
+                'brochure'=>$brochure,
+                'boisson' => $boisson,
+                'Boisson' => $modifierBoissonForm->createView(),
+            ]);
         }
-
-        $brochure =$boisson->getBrochureFilename();
-        $ID = $boisson->getId();
-        return $this->render('Boisson/editBoisson.html.twig', [
-            'ID'=>$ID,
-            'brochure'=>$brochure,
-            'boisson' => $boisson,
-            'Boisson' => $modifierBoissonForm->createView(),
-        ]);
     }
 
     ///////////////////////////////////////////////...Delete...\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
     #[Route('{id}/DeleteBoisson', name: 'Boisson_delete', methods: ['POST'])]
     public function delete(Request $request, Boisson $boisson, BoissonRepository $boissonRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$boisson->getId(), $request->request->get('_token'))) {
-            $boissonRepository->remove($boisson, true);
+        if (!$this->isGranted('IS_AUTHENTICATED_FULLY')) {
+            return $this->redirectToRoute('app_accueil');
+        } else {
+            if ($this->isCsrfTokenValid('delete'.$boisson->getId(), $request->request->get('_token'))) {
+                $boissonRepository->remove($boisson, true);
+            }
+            return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
         }
-        return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
     }
-
 }
