+++ src/Controller/DessertController.php
@@ -20,28 +20,33 @@ class DessertController extends AbstractController
     #[Route('/newDessert', name: 'Dessert_new', methods: ['GET', 'POST'])]
     public function new(Request $request, DessertRepository $DessertRepository, FileUploader $fileUploader): Response
     {
-        $Dessert = new Dessert();
-        $createDessertform = $this->createForm(DessertType::class, $Dessert);
-        $createDessertform->handleRequest($request);
+        if (!$this->isGranted('IS_AUTHENTICATED_FULLY')) {
+            return $this->redirectToRoute('app_accueil');
+        } else {
 
-        if ($createDessertform->isSubmitted() && $createDessertform->isValid()) {
+            $Dessert = new Dessert();
+            $createDessertform = $this->createForm(DessertType::class, $Dessert);
+            $createDessertform->handleRequest($request);
 
-            //################# image ###################\\
-            /** @var UploadedFile $brochureFile */
-            $brochureFile = $createDessertform->get('brochure')->getData();
-            if ($brochureFile) {
-                $brochureFileName = $fileUploader->upload($brochureFile);
-                $Dessert->setBrochureFilename($brochureFileName);
+            if ($createDessertform->isSubmitted() && $createDessertform->isValid()) {
+
+                //################# image ###################\\
+                /** @var UploadedFile $brochureFile */
+                $brochureFile = $createDessertform->get('brochure')->getData();
+                if ($brochureFile) {
+                    $brochureFileName = $fileUploader->upload($brochureFile);
+                    $Dessert->setBrochureFilename($brochureFileName);
+                }
+
+                $DessertRepository->add($Dessert, true);
+                return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
             }
 
-            $DessertRepository->add($Dessert, true);
-            return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
+            return $this->render('Dessert/newDessert.html.twig', [
+                'dessert' => $Dessert,
+                'Dessert' => $createDessertform->createView(),
+            ]);
         }
-
-        return $this->render('Dessert/newDessert.html.twig', [
-            'dessert' => $Dessert,
-            'Dessert' => $createDessertform->createView(),
-        ]);
     }
 
     /////////////////////////////////////////...Update...\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
@@ -49,39 +54,47 @@ class DessertController extends AbstractController
     #[Route('/{id}/editDessert', name: 'menu_edit_Dessert', methods: ['GET', 'POST'])]
     public function edit(Request $request, Dessert $Dessert, DessertRepository $DessertRepository, FileUploader $fileUploader): Response
     {
-        $modifierDessertForm = $this->createForm(DessertType::class, $Dessert);
-        $modifierDessertForm->handleRequest($request);
+        if (!$this->isGranted('IS_AUTHENTICATED_FULLY')) {
+            return $this->redirectToRoute('app_accueil');
+        } else {
+            $modifierDessertForm = $this->createForm(DessertType::class, $Dessert);
+            $modifierDessertForm->handleRequest($request);
 
-        if ($modifierDessertForm->isSubmitted() && $modifierDessertForm->isValid()) {
+            if ($modifierDessertForm->isSubmitted() && $modifierDessertForm->isValid()) {
 
-            //################# image ###################\\
-            /** @var UploadedFile $brochureFile */
-            $brochureFile = $modifierDessertForm->get('brochure')->getData();
-            if ($brochureFile) {
-                $brochureFileName = $fileUploader->upload($brochureFile);
-                $Dessert->setBrochureFilename($brochureFileName);
+                //################# image ###################\\
+                /** @var UploadedFile $brochureFile */
+                $brochureFile = $modifierDessertForm->get('brochure')->getData();
+                if ($brochureFile) {
+                    $brochureFileName = $fileUploader->upload($brochureFile);
+                    $Dessert->setBrochureFilename($brochureFileName);
+                }
+
+                $DessertRepository->add($Dessert, true);
+                return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
             }
-
-            $DessertRepository->add($Dessert, true);
-            return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
+            $brochure =$Dessert->getBrochureFilename();
+            $ID = $Dessert->getId();
+            return $this->render('Dessert/editDessert.html.twig', [
+                'ID'=>$ID,
+                'brochure'=>$brochure,
+                'dessert' => $Dessert,
+                'Dessert' => $modifierDessertForm->createView(),
+            ]);
         }
-        $brochure =$Dessert->getBrochureFilename();
-        $ID = $Dessert->getId();
-        return $this->render('Dessert/editDessert.html.twig', [
-            'ID'=>$ID,
-            'brochure'=>$brochure,
-            'dessert' => $Dessert,
-            'Dessert' => $modifierDessertForm->createView(),
-        ]);
     }
 
     ///////////////////////////////////////////////...Delete...\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
     #[Route('{id}/DeleteDessert', name: 'Dessert_delete', methods: ['POST'])]
     public function delete(Request $request, Dessert $Dessert, DessertRepository $DessertRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$Dessert->getId(), $request->request->get('_token'))) {
-            $DessertRepository->remove($Dessert, true);
+        if (!$this->isGranted('IS_AUTHENTICATED_FULLY')) {
+            return $this->redirectToRoute('app_accueil');
+        } else {
+            if ($this->isCsrfTokenValid('delete'.$Dessert->getId(), $request->request->get('_token'))) {
+                $DessertRepository->remove($Dessert, true);
+            }
+            return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
         }
-        return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
     }
 }
