+++ src/Controller/PlatController.php
@@ -19,27 +19,31 @@ class PlatController extends AbstractController
     #[Route('/newPlat', name: 'Plat_new', methods: ['GET', 'POST'])]
     public function new(Request $request, PlatRepository $PlateRepository, FileUploader $fileUploader): Response
     {
-        $Plat = new Plat();
-        $createPlatform = $this->createForm(PlatType::class, $Plat);
-        $createPlatform->handleRequest($request);
+        if (!$this->isGranted('IS_AUTHENTICATED_FULLY')) {
+            return $this->redirectToRoute('app_accueil');
+        } else {
+            $Plat = new Plat();
+            $createPlatform = $this->createForm(PlatType::class, $Plat);
+            $createPlatform->handleRequest($request);
 
-        if ($createPlatform->isSubmitted() && $createPlatform->isValid()) {
+            if ($createPlatform->isSubmitted() && $createPlatform->isValid()) {
 
-            //################# image ###################\\
-            /** @var UploadedFile $brochureFile */
-            $brochureFile = $createPlatform->get('brochure')->getData();
-            if ($brochureFile) {
-                $brochureFileName = $fileUploader->upload($brochureFile);
-                $Plat->setBrochureFilename($brochureFileName);
+                //################# image ###################\\
+                /** @var UploadedFile $brochureFile */
+                $brochureFile = $createPlatform->get('brochure')->getData();
+                if ($brochureFile) {
+                    $brochureFileName = $fileUploader->upload($brochureFile);
+                    $Plat->setBrochureFilename($brochureFileName);
+                }
+
+                $PlateRepository->add($Plat, true);
+                return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
             }
-
-            $PlateRepository->add($Plat, true);
-            return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
+            return $this->render('Plat/newPlat.html.twig', [
+                'plat' => $Plat,
+                'Plat' => $createPlatform->createView(),
+            ]);
         }
-        return $this->render('Plat/newPlat.html.twig', [
-            'plat' => $Plat,
-            'Plat' => $createPlatform->createView(),
-        ]);
     }
 
 
@@ -48,41 +52,48 @@ class PlatController extends AbstractController
     #[Route('/{id}/editPlat', name: 'menu_edit_plat', methods: ['GET', 'POST'])]
     public function edit(Request $request, Plat $plat, PlatRepository $platRepository, FileUploader $fileUploader): Response
     {
-        $modifierPlatForm = $this->createForm(PlatType::class, $plat);
-        $modifierPlatForm->handleRequest($request);
+        if (!$this->isGranted('IS_AUTHENTICATED_FULLY')) {
+            return $this->redirectToRoute('app_accueil');
+        } else {
+            $modifierPlatForm = $this->createForm(PlatType::class, $plat);
+            $modifierPlatForm->handleRequest($request);
 
-        if ($modifierPlatForm->isSubmitted() && $modifierPlatForm->isValid()) {
+            if ($modifierPlatForm->isSubmitted() && $modifierPlatForm->isValid()) {
 
-            //################# image ###################\\
-            /** @var UploadedFile $brochureFile */
-            $brochureFile = $modifierPlatForm->get('brochure')->getData();
-            if ($brochureFile) {
-                $brochureFileName = $fileUploader->upload($brochureFile);
-                $plat->setBrochureFilename($brochureFileName);
+                //################# image ###################\\
+                /** @var UploadedFile $brochureFile */
+                $brochureFile = $modifierPlatForm->get('brochure')->getData();
+                if ($brochureFile) {
+                    $brochureFileName = $fileUploader->upload($brochureFile);
+                    $plat->setBrochureFilename($brochureFileName);
+                }
+
+                $platRepository->add($plat, true);
+                return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
             }
 
-            $platRepository->add($plat, true);
-            return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
+            $brochure =$plat->getBrochureFilename();
+            $ID = $plat->getId();
+            return $this->render('plat/editPlat.html.twig', [
+                'brochure'=>$brochure,
+                'plat' => $plat,
+                'Plat' => $modifierPlatForm->createView(),
+                'ID'=>$ID,
+            ]);
         }
-
-        $brochure =$plat->getBrochureFilename();
-        $ID = $plat->getId();
-        return $this->render('plat/editPlat.html.twig', [
-            'brochure'=>$brochure,
-            'plat' => $plat,
-            'Plat' => $modifierPlatForm->createView(),
-            'ID'=>$ID,
-        ]);
     }
 
     ///////////////////////////////////////////////...Delete...\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
     #[Route('/{id}/DeletePlat', name: 'Plat_delete', methods: ['post'])]
     public function delete(Request $request, Plat $plat, PlatRepository $PlatRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$plat->getId(), $request->request->get('_token'))) {
-            $PlatRepository->remove($plat, true);
+        if (!$this->isGranted('IS_AUTHENTICATED_FULLY')) {
+            return $this->redirectToRoute('app_accueil');
+        } else {
+            if ($this->isCsrfTokenValid('delete'.$plat->getId(), $request->request->get('_token'))) {
+                $PlatRepository->remove($plat, true);
+            }
+            return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
         }
-        return $this->redirectToRoute('app_menu', [], Response::HTTP_SEE_OTHER);
     }
-
 }
